We don't need to override element_name in Resource
That's a legacy method from the 37Signals wrapper and i'm not sure why it was there. It doesn't seem to affect rails < 3 (the method is ignored) but it affects rails >= 3 when the name of the resource in the API is different than the name of the class, for example Basecamp::Message that should use "post" as the resource name in the API calls.